fix: patternProperties panic, integer width selection, range checks (#882, #975)
#882 — Replace assertion panics in merge.rs for patternProperties and
propertyNames with best-effort pass-through behavior.

#975 — When min and max are both specified, find the smallest integer
type that contains the range (iterate forward instead of reverse).
Integer [1..32] now generates NonZeroU8 instead of NonZeroU64.

Fix range constraint codegen for NonZero types: use .get() to extract
the inner value for comparison instead of invalid `as` cast.

Runtime tests: SmallRange (NonZeroU8) TryFrom and serde validation.

+++ typify/tests/schemas/simple-types.rs
@@ -77,10 +77,10 @@ impl ::std::convert::From<Dscp> for u8 {
 impl ::std::convert::TryFrom<u8> for Dscp {
     type Error = self::error::ConversionError;
     fn try_from(value: u8) -> ::std::result::Result<Self, self::error::ConversionError> {
-        if value < 0i64 as u8 {
+        if (value as i64) < 0i64 {
             return Err("value must be >= 0".into());
         }
-        if value > 63i64 as u8 {
+        if (value as i64) > 63i64 {
             return Err("value must be <= 63".into());
         }
         Ok(Self(value))
